// src/app/interfaces/common/reports.interface.ts
/* eslint-disable prettier/prettier */
export interface Reports {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  title?: string;
  shortDescription?:string;
  description?: string;
  shortDescriptionEn?:string;
  descriptionEn?: string;
  createdAt?: Date;
  updatedAt?: Date;
}